+++ doc/changelog.rst
@@ -15,6 +15,7 @@ Unreleased
 * Fixed: Input URLs are not fetched in order. Regression since 1.1.
 * Fixed: UnicodeEncodeError when fetching FTP files with non-ASCII filenames.
 * Fixed: Session cookies not loaded when using ``--load-cookies``.
+* Fixed: ``--keep-session-cookies`` was always on.
 * Changed: FTP communication uses UTF-8 instead of Latin-1.
 * Changed: ``--prefer-family=none`` is now default.
 * Added: ``none`` as a choice to ``--prefer-family``.